cmdb-backend/requirements.txt
aliyun-python-sdk-cdn==3.7.5
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-core-v3==2.13.33
aliyun-python-sdk-ecs==4.24.26
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
community==1.0.0b1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
fabric==2.7.1
flake8==6.0.
idna==3.3
importlib-metadata==5.0.0
install==1.3.5
invoke==1.7.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
networkx==2.8.8
openapi-codec==1.3.2
paramiko==2.11.0
pathlib2==2.3.7.post1
pipreqs==0.4.11
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==3.0.1
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.5.0
python-louvain==0.16
pytz==2021.3
regex==2022.10.31
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.2.2
xlrd==1.2.0
yarg==0.1.9
zipp==3.10.0
